psblas3:  base/modules/psb_base_tools_mod.f90  base/tools/psb_cdins.f90  docs/html/footnode.html  docs/html/img100.png  docs/html/img101.png  docs/html/img102.png  docs/html/img103.png  docs/html/img104.png  docs/html/img105.png  docs/html/img106.png  docs/html/img107.png  docs/html/img108.png  docs/html/img109.png  docs/html/img110.png  docs/html/img111.png  docs/html/img112.png  docs/html/img113.png  docs/html/img114.png  docs/html/img115.png  docs/html/img116.png  docs/html/img117.png  docs/html/img118.png  docs/html/img119.png  docs/html/img120.png  docs/html/img121.png  docs/html/img122.png  docs/html/img123.png  docs/html/img124.png  docs/html/img125.png  docs/html/img126.png  docs/html/img127.png  docs/html/img128.png  docs/html/img129.png  docs/html/img130.png  docs/html/img131.png  docs/html/img132.png  docs/html/img133.png  docs/html/img134.png  docs/html/img135.png  docs/html/img136.png  docs/html/img137.png  docs/html/img138.png  docs/html/img139.png  docs/html/img140.png  docs/html/img141.png  docs/html/img142.png  docs/html/img143.png  docs/html/img144.png  docs/html/img145.png  docs/html/img146.png  docs/html/img147.png  docs/html/img148.png  docs/html/img149.png  docs/html/img19.png  docs/html/img20.png  docs/html/img21.png  docs/html/img22.png  docs/html/img23.png  docs/html/img24.png  docs/html/img25.png  docs/html/img26.png  docs/html/img27.png  docs/html/img28.png  docs/html/img29.png  docs/html/img30.png  docs/html/img31.png  docs/html/img32.png  docs/html/img33.png  docs/html/img34.png  docs/html/img35.png  docs/html/img36.png  docs/html/img37.png  docs/html/img38.png  docs/html/img39.png  docs/html/img40.png  docs/html/img41.png  docs/html/img42.png  docs/html/img43.png  docs/html/img44.png  docs/html/img45.png  docs/html/img46.png  docs/html/img47.png  docs/html/img48.png  docs/html/img49.png  docs/html/img50.png  docs/html/img51.png  docs/html/img52.png  docs/html/img53.png  docs/html/img54.png  docs/html/img55.png  docs/html/img56.png  docs/html/img58.png  docs/html/img59.png  docs/html/img60.png  docs/html/img61.png  docs/html/img62.png  docs/html/img63.png  docs/html/img64.png  docs/html/img65.png  docs/html/img66.png  docs/html/img67.png  docs/html/img68.png  docs/html/img69.png  docs/html/img70.png  docs/html/img71.png  docs/html/img72.png  docs/html/img73.png  docs/html/img74.png  docs/html/img75.png  docs/html/img76.png  docs/html/img77.png  docs/html/img78.png  docs/html/img79.png  docs/html/img80.png  docs/html/img81.png  docs/html/img82.png  docs/html/img83.png  docs/html/img84.png  docs/html/img85.png  docs/html/img86.png  docs/html/img87.png  docs/html/img88.png  docs/html/img89.png  docs/html/img90.png  docs/html/img91.png  docs/html/img92.png  docs/html/img93.png  docs/html/img94.png  docs/html/img95.png  docs/html/img96.png  docs/html/img97.png  docs/html/img98.png  docs/html/img99.png  docs/html/index.html  docs/html/node10.html  docs/html/node100.html  docs/html/node101.html  docs/html/node113.html  docs/html/node118.html  docs/html/node120.html  docs/html/node20.html  docs/html/node3.html  docs/html/node30.html  docs/html/node36.html  docs/html/node38.html  docs/html/node39.html  docs/html/node4.html  docs/html/node40.html  docs/html/node41.html  docs/html/node42.html  docs/html/node43.html  docs/html/node44.html  docs/html/node45.html  docs/html/node46.html  docs/html/node47.html  docs/html/node48.html  docs/html/node49.html  docs/html/node50.html  docs/html/node52.html  docs/html/node53.html  docs/html/node54.html  docs/html/node55.html  docs/html/node57.html  docs/html/node58.html  docs/html/node62.html  docs/html/node63.html  docs/html/node64.html  docs/html/node68.html  docs/html/node69.html  docs/html/node7.html  docs/html/node70.html  docs/html/node72.html  docs/html/node73.html  docs/html/node74.html  docs/html/node75.html  docs/html/node76.html  docs/html/node77.html  docs/html/node78.html  docs/html/node8.html  docs/html/node81.html  docs/html/node83.html  docs/html/node85.html  docs/html/node86.html  docs/html/node87.html  docs/html/node89.html  docs/html/node93.html  docs/html/node94.html  docs/html/node95.html  docs/html/node96.html  docs/html/node97.html  docs/html/node98.html  docs/html/node99.html  docs/html/userhtml.html  docs/psblas-3.0.pdf  docs/src/intro.tex  docs/src/toolsrout.tex  docs/src/userguide.tex  docs/src/userhtml.tex
Towards the "official" 3.0 release: 
1. fixed minor interface problem with LIDX in cdins.
2. Updated docs with  cross-references for usage of LIDX/LOCAL.

+++ base/modules/psb_base_tools_mod.f90
@@ -362,12 +362,12 @@ module psb_cd_if_tools_mod
     end subroutine psb_cdinsrc
     subroutine psb_cdinsc(nz,ja,desc,info,jla,mask,lidx)
       import :: psb_ipk_, psb_desc_type
-      type(psb_desc_type), intent(inout) :: desc
-      integer(psb_ipk_), intent(in)                :: nz,ja(:)
-      integer(psb_ipk_), intent(out)               :: info
-      integer(psb_ipk_), optional, intent(out)     :: jla(:)
-      logical, optional, target, intent(in) :: mask(:)
-      integer, intent(in), optional         :: lidx(:)
+      type(psb_desc_type), intent(inout)         :: desc
+      integer(psb_ipk_), intent(in)              :: nz,ja(:)
+      integer(psb_ipk_), intent(out)             :: info
+      integer(psb_ipk_), optional, intent(out)   :: jla(:)
+      logical, optional, target, intent(in)      :: mask(:)
+      integer(psb_ipk_), intent(in), optional    :: lidx(:)
     end subroutine psb_cdinsc
   end interface
